% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwgs_gilles.R
\name{bwgs.cv}
\alias{bwgs.cv}
\title{Genomic Prediction with cross validation}
\usage{
bwgs.cv(
  geno,
  pheno,
  FIXED = "NULL",
  MAXNA = 0.2,
  MAF = 0.05,
  pop.reduct.method = "NULL",
  sample.pop.size = "NULL",
  geno.reduct.method = "NULL",
  reduct.marker.size = "NULL",
  pval = "NULL",
  r2 = "NULL",
  MAP = "NULL",
  geno.impute.method = "NULL",
  predict.method = "NULL",
  nFolds,
  nTimes
)
}
\arguments{
\item{geno}{A matrix. Genotypes coded with (-1 0 1). Individuals in rows. Markers in column.}

\item{pheno}{A matrix. One column per character.}

\item{FIXED}{A matrix of fixed effect, to be used with some methods such as those included in BGLR, MUST have same rownames as geno and coded(-1 0 1)}

\item{MAXNA}{maximum^proportion of missing data. Markers with NA > this throshold are skipped out}

\item{MAF}{Minor allele frequency for Data cleaning. Markers with MAF < this thersholds are skipped out. Can generate errors when unadapted}

\item{sample.pop.size}{A number < nrow(geno)}

\item{geno.reduct.method}{A character string. It can be "RMR", "ANO", "LD","ANO+LD" or "NULL". See documentation}

\item{pval}{: the Pvalue threshold to be used for marker selection by independent GWAS.}

\item{r2}{A number. If geno.reduct.method="LD" or geno.reduct.method="ANO+LD" .}

\item{MAP}{A matrix with at least one column entitled "chrom" with chromosome number (possibly character)}

\item{geno.impute.method}{A character string. It can be "MNI" or "EMI".}

\item{predict.method}{A character string. It can be "GBLUP","EGBLUP","RR","LASSO","EN"n"BRR","BL","BA","BB","BC","RKHS","MKRKFS","RF","SVM","BRNN".}

\item{nFolds}{A number. The number of partitions in the data for cross validation. If 5, 4/5 of the data will be training data and 1/5 validation data.}

\item{nTimes}{A number. The number of times to randomly repeat cross validation.}

\item{pop.reduct.method:}{a character string, option to select individuals for further analyses. options are "RANDOM" or "OPTI", based on the optimization algorithm of Rincent et al 2012}

\item{reduct.marker.size:}{the number of marker to be sampled or selected.}
}
\description{
Calculates GEBV using different models and cross validation. See Documentation for more details.
}
\examples{
bwgs.cv(geno,pheno,random.pop.size="NULL",geno.reduct.method="NULL",reduct.size=NULL,pval=NULL,
}
