% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwgs_gilles.R
\name{bwgs.predict}
\alias{bwgs.predict}
\title{Title}
\usage{
bwgs.predict(
  geno_train,
  pheno_train,
  geno_target,
  FIXED_train = "NULL",
  FIXED_target = "NULL",
  MAXNA = 0.2,
  MAF = 0.05,
  geno.reduct.method = "NULL",
  reduct.size = "NULL",
  r2 = "NULL",
  pval = "NULL",
  MAP = "NULL",
  geno.impute.method = "NULL",
  predict.method = "GBLUP"
)
}
\arguments{
\item{geno_train}{}

\item{pheno_train}{}

\item{geno_target}{}

\item{FIXED_train}{}

\item{FIXED_target}{}

\item{MAXNA}{}

\item{MAF}{}

\item{geno.reduct.method}{}

\item{reduct.size}{}

\item{r2}{}

\item{pval}{}

\item{MAP}{}

\item{geno.impute.method}{}

\item{predict.method}{}
}
\value{

}
\description{
Title
}
