% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inra47k.R
\docType{data}
\name{inra47k}
\alias{inra47k}
\alias{MAP47K}
\alias{TARGET47K}
\alias{TRAIN47K}
\alias{YieldBLUE}
\title{INRA47K}
\format{
An object of class \code{data.frame} with 10000 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 10000 columns.

An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 10000 columns.

An object of class \code{numeric} of length 100.
}
\source{
<https://forgemia.inra.fr/umr-gdec/bwgs>
}
\usage{
MAP47K

TARGET47K

TRAIN47K

YieldBLUE
}
\description{
inra data contains a set of geno47K(760 x 47839), pheno (760 x 1) and MAP47K (47839 x 3). The phenotype pheno contains adjusted genotype means for yield trait (YLD) over multi-year/location trials.
}
\keyword{datasets}
