% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwgs_gilles.R
\name{removeSaveAt}
\alias{removeSaveAt}
\title{Remove saveAt file generated by BGLR package}
\usage{
removeSaveAt(saveAt)
}
\arguments{
\item{saveAt}{saveAt}
}
\description{
Remove saveAt file generated by BGLR package
}
